+++ compiler-plugin/src/main/kotlin/tech/mappie/exceptions/MappiePanicException.kt
@@ -1,0 +1,5 @@
+package tech.mappie.exceptions
+
+import org.jetbrains.kotlin.ir.IrElement
+
+class MappiePanicException(message: String, val origin: IrElement? = null) : Exception(message)