+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/objects/GeneratedClassListTest.kt
@@ -1,0 +1,51 @@
+package tech.mappie.testing.objects
+
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.io.TempDir
+import tech.mappie.testing.compilation.KotlinCompilation
+import tech.mappie.testing.compilation.KotlinCompilation.ExitCode
+import tech.mappie.testing.compilation.SourceFile.Companion.kotlin
+import tech.mappie.testing.loadObjectMappieClass
+import java.io.File
+
+class GeneratedClassListTest {
+    data class Input(val a: InnerInput)
+    data class InnerInput(val value: List<String>)
+    data class Output(val a: InnerOutput)
+    data class InnerOutput(val value: List<String>)
+
+    @TempDir
+    lateinit var directory: File
+
+    @Test
+    fun `map data classes with nested list without declaring mapping should succeed`() {
+        KotlinCompilation(directory).apply {
+            sources = buildList {
+                add(
+                    kotlin("Test.kt",
+                        """
+                        import tech.mappie.api.ObjectMappie
+                        import tech.mappie.testing.objects.GeneratedClassListTest.*
+    
+                        class Mapper : ObjectMappie<Input, Output>()
+                        """
+                    )
+                )
+            }
+        }.compile {
+            assertThat(exitCode).isEqualTo(ExitCode.OK)
+            assertThat(messages).isEmpty()
+
+
+            val mapper = classLoader
+                .loadObjectMappieClass<Input, Output>("Mapper")
+                .constructors
+                .first()
+                .call()
+
+            assertThat(mapper.map(Input(InnerInput(listOf("first", "second")))))
+                .isEqualTo(Output(InnerOutput(listOf("first", "second"))))
+        }
+    }
+}